Rename files to match conventions
